Refactoring code and url handling
Refactored code such that when we input a superset url in the superset url field on the login page, that doesnt break our links. As in the previous version, we were passing it as a parameter in the url. Now this link is passed around via state calls and functions similar to props in normal JSX components.

TODO: figure out why useNavigate isnt working upon a successful clone